Working async registration
