+++ packages/common/src/auth/index.ts
@@ -2,3 +2,4 @@ export { AuthModule } from './auth.module';
 export { JwtAuthGuard } from './jwt.guard';
 export { TokenContext } from './token.context';
 export { TokenPayload } from './token.dto';
+export { AuthModuleOptions, AuthModuleOptionsAsync } from './auth.interface';
