+++ packages/common/src/auth/jwt.strategy.ts
@@ -1,14 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Inject } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
+import { JWT_MODULE_OPTIONS } from '@nestjs/jwt/dist/jwt.constants';
+import { JwtModuleOptions } from '@nestjs/jwt';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor(publicKey: string) {
+  constructor(@Inject(JWT_MODULE_OPTIONS) jwtOptions: JwtModuleOptions) {
     super({
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
       ignoreExpiration: false,
-      secretOrKey: publicKey
+      secretOrKey: jwtOptions.publicKey!
     });
   }
